requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-taggit==1.5.1
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
uWSGI==2.0.20
